Prepare for final Simulation

+++ Algorithm/PreAlgo.m
@@ -188,8 +188,8 @@ if CostCats(1)
 end
 
 if ismember(TimeInd, TimesOfPreAlgo(1,:))
-    ChargingMat{1}(:,:,:,PreAlgoCounter)=OptimalChargingEnergies;
-    AvailabilityMat=[AvailabilityMat, Availability(1:24*Time.StepInd,1,:)];
+    ChargingMat{1}(:,:,:,PreAlgoCounter)=single(OptimalChargingEnergies);
+    AvailabilityMat=single([AvailabilityMat, Availability(1:24*Time.StepInd,1,:)]);
     
     SuccessfulResPoOffers(:,PreAlgoCounter+1)=ResPoOffers(:,1,PreAlgoCounter+1)<=repelem(ResPoPricesReal4H(floor((TimeInd+TD.Main)/(4*Time.StepInd))+1+(24-hour(TimeOfPreAlgo(1)))/4:floor((TimeInd+TD.Main)/(4*Time.StepInd))+(24-hour(TimeOfPreAlgo(1)))/4+6,5)/1000, ConstantResPoPowerPeriodsScaling); %[EUR/kW]
     LastResPoOffers(:,PreAlgoCounter+1)=sum(OptimalChargingEnergies(1:ConstantResPoPowerPeriods:end,sum(CostCats(1:3)),:), 3); % [Wh]
